+++ src/program/Config.cpp
@@ -69,6 +69,7 @@ void Config::save(const std::string& gamepath) {
     general_settings.endArray();
 
     general_settings.setValue("debugger", debugger);
+    general_settings.setValue("allow_downloads", allow_downloads);
 
     /* Open the preferences for the game */
     QSettings settings(iniPath(gamepath), QSettings::IniFormat);
@@ -218,6 +219,7 @@ void Config::load(const std::string& gamepath) {
 #elif defined(__APPLE__) && defined(__MACH__)
     debugger = DEBUGGER_LLDB;
 #endif
+    allow_downloads = general_settings.value("allow_downloads", allow_downloads).toInt();
 
     ffmpegoptions = general_settings.value("ffmpegoptions", ffmpegoptions.c_str()).toString().toStdString();